refactor with four arch
