+++ src/bare/arch/aarch64/page_table.rs
@@ -1,7 +1,7 @@
 use aarch64_cpu::registers::{Writeable, TTBR0_EL1};
 
-use crate::addr::{PhysAddr, PhysPage, VirtAddr, VirtPage};
-use crate::pagetable::{MappingFlags, PageTable, PTE, TLB};
+use crate::{PhysAddr, PhysPage, VirtAddr, VirtPage};
+use crate::{MappingFlags, PageTable, PTE, TLB};
 
 impl PTE {
     #[inline]
